Update README.md
NASA releases the Astronomy Picture of the Day everyday. This is accompanied by the title of the picture and a short explanation of about it.
We want to build a iOS/Android App where the user can see the daily picture.
Consider this page as a small part of a large App and structure it accordingly. Use the architecture of your choice that would work best for such an App.
API source: https://api.nasa.gov API key required: Yes Endpoint: https://api.nasa.gov/planetary/apod UX: Use your judgement on best practices, be conscious of the time limit and, keep the UI simple. There are no extra points for very fancy UI.